requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.3.0
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
phonenumbers==8.12.4
cython
gunicorn
gevent
